StringCompression.py - list comp & range()
This solution uses a for loop that demonstrates how to use range diversely, and properly use it if you want to access the previous item in the loop. 
It also utilizes list comprehensions, a more advanced technique in Python along with join() -- showing necessary knowledge for when working with Python.