Update RotrateMatrix.py - shorter, zip, range()
This utilizes zip() in order to simultaneously access the indices increasing and decreasing to properly assign values. 
This also demonstrates understanding of list comprehensions, and how []*n works, since it cannot be lazily written as [[]*n]*n for assigning different values in this case. 
Also, I added another test case -- a more simple one so when introducing the problem you can visualize it a little easier.