Perf best practice in isUnique.py
Not to be stickler or over-optimizer, but in the spirit of foster best practices of Python through education material I have to say according to my benchmark using `timeit` module, `[False] * number` in 3 orders of magnitude faster even in small number such as 128, to general best practice is to use this instead of a for loop to initialize an array with fix value. 
And doing so in great educational materials like yours would hopefully make people more aware of such good syntaxes.