Clarification question about directed graphs.
4.1 Route Between Nodes: "Given the 'directed' graph, design an algorithm to find out whether there is a route between two nodes."

My DFS implementation was very similar to the one shown above; however, I kept getting a keyword error for one of the nodes when I tried to make a recursive call on the leaf nodes. I tried to reproduce the error using the code above, and received the same error when using the test graph I had designed, but not when using the graph diagrammed above. 

From my understanding of the explanation in the book, an undirected graph is similar to a two-way street, while a directed graph is like a one-way street. That would make it seem like the above implementation is not a directed graph (unless I misunderstood what a directed graph is). I have edited the changes that would make the above a graph into what I understand to be a directed graph. Running the above code should generate the following error ("O" and "L" are also one-way street):

Traceback (most recent call last):
  File "c:\Users\Cleme\Practice_Code\Interview_prep\CCI_6ed\Chapter_4\route_between_nodes_answer.py", line 65, in test_is_route
    actual = is_route(self.graph, start, end)
  File "c:\Users\Cleme\Practice_Code\Interview_prep\CCI_6ed\Chapter_4\route_between_nodes_answer.py", line 27, in is_route
    if node == end or is_route(graph, node, end, visited):
  File "c:\Users\Cleme\Practice_Code\Interview_prep\CCI_6ed\Chapter_4\route_between_nodes_answer.py", line 27, in is_route
    if node == end or is_route(graph, node, end, visited):
  File "c:\Users\Cleme\Practice_Code\Interview_prep\CCI_6ed\Chapter_4\route_between_nodes_answer.py", line 27, in is_route
    if node == end or is_route(graph, node, end, visited):
  [Previous line repeated 3 more times]
  File "c:\Users\Cleme\Practice_Code\Interview_prep\CCI_6ed\Chapter_4\route_between_nodes_answer.py", line 24, in is_route
    for node in graph[start]:
KeyError: 'O'

I might be overanalyzing the problem, but from my understanding from the book and from looking on the web, the above implementation is not a directed graph.